Temporarily restore recursive behaviour for bundle filesets (#4624)
### Problem

The `fileset` argument to `jvm_app` `bundles` is approximately like the `sources` argument on a target, with the exception that if a directory is matched, its contents are implicitly recursively included. Usually triggering recursive inclusion requires `**/*` wildcarding or `rglobs`.

Additionally, the v2 engine broke the recursive inclusion by excluding directories from the `FilesetWithSpec` that was created in the case of a `bundle`.

### Solution

Restore capturing directories for `FilesetWithSpec`s (but only in the context of bundles, for now), and deprecate the inconsistent default-recursive behaviour of the `fileset` argument.

### Result

Users are encouraged to switch to explicit recursive globs, which will allow us to remove the inconsistent behavior of the `fileset` argument. Fixes #4622.