Rename and expand usage of the needs-cherrypick label (#4652)
### Problem

The naming of the `needs-rc-cherrypick` label would seem to imply that cherrypicks only happen for `rcs`, but the label is useful anytime something should be cherrypicked to a stable branch. Also, we've begun to use milestones more aggressively, but currently they are not mentioned on the Release Strategy page. 

### Solution

Rename the label to `needs-cherrypick`, and include a link to github milestones. Also, remove an outdated (/wishful?) reference to a "two week" timeframe.