Skip a test that assumes the current version is a pre-release version. (#4654)
### Problem

Sometime during the `1.3.0.dev` cycle we added validation that deprecation versions are `dev`/prerelease versions (to avoid having a bunch of deprecations fire the instant a stable release was cut). But one test was not updated for this validation, and so bumping the version to `1.3.0` in the stable branch caused it to fail.

### Solution

Skip that test if the assumption (that the current version is a prerelease version) does not hold. Another approach would have been to add arguments to the deprecation package to allow for spoofing the current version, but that would have pretty wide-ranging impacts for something that can be easily ignored.