Add compiler_args property to JavaThriftLibrary target.  (#4669)
### Problem

For supporting the generation of additional scrooge code types, flags might need to be passed to the scrooge compiler.

### Solution

Instead of adding another special option to JavaThriftLibrary target, add an option to specify any compiler arguments. Additionally, deprecate the `rpc_style` argument, as it is easily subsumed by `compile_args`.

### Result

Users will now be able to specify any compiler arguments that they need to pass to either scrooge or the apache thrift compiler.