Remove engine.engine.Engine and add RootRule (#4679)
### Problem

As described on #4229, the `engine.engine.Engine` class is no longer necessary (as execution is handled entirely in the native code), and introduces unnecessary complexity into executing requests for products.

Additionally, the installation of "root product types" (which are used to describe locations in the rule graph where an execution request might provide a type that isn't otherwise possible to compute with installed Rules) was disconnected from the installation of the rules themselves. This made testing awkward in many situations, because there was a default set of root product types but tests frequently wanted to use another set that was specific to the rules that they were installing.

### Solution

Remove `engine.engine.Engine` and move its two interesting methods to `engine.scheduler.Scheduler`. "Engine" was a pretty good name, so maybe we can rename "Scheduler" at some point. But right now it seemed like it would confuse things.

Additionally, added `RootRule` to unify the declaration of providers for types under the `Rule` interface, and allow a set of Rules to declare its own roots directly.

### Result

Testing and usage is cleaner. Fixes #4229.