Ensure that target root order is preserved (#4708)
The 1.3.0 stable branch contains an issue where target root order is not always preserved, which leads to `./pants paths $X $Y` looking for paths from `$Y` to `$X` ~50% of the time. Apparently `paths` is (one of?) the only tasks that is order-dependent with regard to target roots. There are unit tests to validate that roots are ordered correctly, but due to #4401 they are not all providing coverage yet.

This bug does not actually exist in master (because it was fixed in #4679), but while looking for the bug I noticed that the relevant code could use further cleanup, which will be cherry-picked to the `1.3.0` branch.

Add integration test coverage for `paths`, with a comment to remove it in favor of the unit tests once #4401 lands.