Ensure that invalidation works correctly when state is reverted.
Previously, a failing task could make changes to the workdir,
but wouldn't have its invalidation key recorded, so reverting to
the input state at the previous successful build would make it look
as if the output state was valid when it wasn't.

We used to special-case this in JvmCompile. Now we solve this
globally.