Fixup `PytestRun` error handling. (#4716)
Previously the task could trip up postprocessing an execution error by
blindly expecting a junit xml file to have been created by the test
run. Fix things so that the existence of the junit xml file is the
sentinel that a test run was executed and fail fast if the sentinel does
not exist. This also fixes a latent bug where prior test failures could
pullute subsequent failed runs.

Fixes #4712