Extract a zinc subsystem to allow for more entrypoints (#4720)
### Problem

Pants' zinc wrapper tool is currently configured directly on the `compile.zinc` task, which would cause duplication if some other task wanted to use the zinc install (see #4477 in particular). 

### Solution

Extract a `zinc` subsystem which is required by the `compile.zinc` task, and which can be consumed from other tasks in the future.

### Result

Because the tool key is the same, only the options which allow for re-configuring the tool location needed to be deprecated: if they were overriding tool locations, users will need to move them from `--compile-zinc-zinc` to `--zinc-zinc`, etc.