Improve pytest result summaries.
This fixes the previously confusing output for `--fast` with failures.
Revert to providing no summary output in this case, since we can't
identify the failing targets without introducing code to parse the junit
xml from the run and then associate failures to files to target owners.

The summaries are still displayed however for `--fast` success. This
provides feedback in the "cached" test success run case.

 Fixes #4736