Move support for `sources` into `Target`.
There were several target subclasses re-implementing the boilerplate and
`Target` otherwise had intimiate knowledge of sources and an expectation
that the field existed for auxially methods.

This also allows for a general `sources` bag that can be used to express
file dependencies with no other expectation about the files' type or
purpose.