Introduce a loose `Files` target. (#4798)
This will be used initially to allow test targets to declare
dependencies on loose files that should be included in test chroots. An
example of this use is included which lists the correct `filedeps` for
an existing integration test. Follow-up changes will add support for
processing these targets to `PytestRun` and `JunitRun`.