Clean up stray pantsd-runner processes (#4835)
Problem

There are stray pantsd-runners that continue to live on even after a run has finished when an exception is thrown in pants. The error is detailed in issue 4827.

Solution

To ensure we are exiting out of the process when there is an exception, wrap the os._exit in a finally block.

Result

With this change, the pantsd-runner process closes and we don't have any stray processes running.