Downgrade to vanilla requests.
requests[security] has a dependency on pyopenssl, which is currently
triggering a depreaction due to an OpenSSL.rand call.

pyopenssl has fixed this issue in code, but is waiting for a maintainer
to return from vacation and cut a release.

In the meantime, the release cannot be completed with the current
dep due to the non-zero it raises.

      .7/site-packages/urllib3/contrib/pyopenssl.py:46: DeprecationWarning: OpenSSL.rand is deprecated - you should use os.urandom instead
  import OpenSSL.SSL
      **** Failed to install cryptography-2.0.3 (caused by: NonZeroExit("received exit code 1 during execution of

This was a CI-fix around ssl handshakes, and since the CI is currently having its own
issues, I figured we could unlock the release while we wait.

If this commit lands, it should be short-term and be undone once the
pyopenssl can be released.

References: #4856