Move to SymbolTable/Parser instances (#4864)
### Problem

Currently the `SymbolTable` and `Parser` used in v2 engine BUILD file parsing are static, and access the `BuildFileAliases` that they use by lazily triggering configuration parsing on use. This detracts from testability, and amounts to a "nasty escape hatch" that would have needed to be propagated in #4401.

Part of the reason these were made static long ago was that pickling was used to compute object identities, and to support multiprocessing of all objects manipulated by the engine. Things have changed significantly since that time:
1. object identities need not be stable across processes, since they will never be used in cache keys (only process invokes will be cached)
2. multiprocessing for individual function invokes is no longer a goal, and so universal pickleability is no longer a goal. If we multiprocess, it will be at a much larger scope (ie, starting from pickleable inputs, redundantly compute BuildFileAliases, for example). 

### Solution

1. Remove the requirement for pickleability of objects manipulated in the engine by switching back to memoization for the computation of object ids.
2. Switch to using instances of the SymbolTable and Parser, removing the requirement for static initialization
3. Drop the `storage.py` module based on the above understanding of content-addressability/caching.

### Result

The nasty escape hatch is removed, and `setup_legacy_graph` accepts an explicit `BuildFileAliases` object to construct a `SymbolTable` from. Followup in #4401 will continue porting BaseTest to the v2 engine.