Reduce BUILD file parse pollution (#4892)
### Problem

As described on #4880, the parser used in the v2 engine was re-using a symbol table across parses, which (in the presence of non-deterministic execution orders due to engine concurrency) could cause non-deterministic successful parses of otherwise invalid code.

### Solution

Shallow clone the symbol table for each use, which should defend against 95% of errors in this area. The remaining 5% would require deep cloning (or recreating) the symbol table for each file parse, which doesn't seem worth the benefit.

### Result

A failing test that imports in parsed files now succeeds, and non-determinism due to imports should be eliminated. Fixes #4880