Improve perf of simple codegen.
There are two related issues fixed here. The first is to batch the
invalidation of dependees of injected synthetic targets, which reduces
the number of times we need to walk the build graph. The second is to
mark the dependees of the synthetic target dirty, rather than the
dependees of the dependencies of the synthetic target, which reduces the
number of targets that are invalidated.