Refactor code coverage in preparation for adding a new coverage engine (jacoco) (#4881)
### Problem

There is tight coupling between the Cobertura coverage class and junit_run that makes it difficult to add a new coverage system. The use of an abstract BaseClass makes sharing common coverage parameters difficult.

### Solution

Refactoring to eliminate abstract BaseCoverage class, in favor of a utility class and a class to manage coverage. The manager can be turned into a subsystem, and the coverage classes into dependent subsystems down the road. Not doing that now because it would impact the parameter scoping, so requires a deprecation cycle.

### Result

After the refactor, added a new, no-op coverage system for jacoco. In a subsequent branch, will implement jacoco coverage inside of this class.