Move `buildozer` to the contrib module
This moves the `buildozer` goal to the contrib module and adds it as a plugin to
pants. Note that `buildozer` should now show up as a goal in `./pants goals`.

This also removes the previous behavior of `location` and functions based of
`self.context.target_roots` instead.

Finally, select_binary is used over select_script since the `buildozer` goal
isn't platform depenendent.