Memoize strict deps and exports (#4934)
### Problem

Computation of both `strict_dependencies` and `exports` are occasionally mutually recursive: `strict_dependencies` uses `exports`, which can call back into `strict_dependencies`, and calls to `exports` are transitive with other calls to `exports`, in general.

This leads to performance bottlenecks in the computation of `strict_dependencies` where we repeatedly re-compute both fields. This is particularly slow for `exports`, which involve parsing `Addresses` from specs and looking them up in the `BuildGraph`.

### Solution

Memoize the `strict_dependencies` and (direct) `exports` of a Target.

While the memoization that `Target` does is not ideal (it's very adhoc, and requires explicit invalidation when the BuildGraph changes), memoizing the `strict_dependencies`/`exports` for a `Target` anywhere else would likely be even more error prone. 

### Result

The overhead of `main:compile:zinc` relative to `main:compile:zinc:compile:zinc` (ie, the time spent inside the `compile.zinc` pants task, relative to time spent in the compiler) is reduced by 20% for strict dependencies targets, such that with `strict_dependencies` enabled (versus disabled), we spend less time in the compiler, _and_ less time time in dependency calculation.