Preserve soft excludes bug while removing duplicates (#4940)
### Problem

`soft_excludes` have been broken at compile time for a while: see #4874 and the comments here.

### Solution

This pull exposes the reason for #4874, and adds an xfailed test. Unfortunately, actually fixing this issue is going to require some fundamental API changes: we no longer walk the `BuildGraph` while computing the classpath: instead we walk it externally. And `excludes` are graph position specific. 

### Result

#4874 is marked, and the current behaviour is preserved, for now.