Pin Rust version to 1.20.0
This gives better reproducability, both for different users, and for
historical commits.

Re-sets the override every build.

Also includes the native build-support files in the
native_engine_version, so that changes to how we build the engine are
reflected in its hash.

Also also, redirect some processes' stdout to stderr because the stdout
of this function is used in bootstrap_native_code.