Allow users to tell pants where to look for python interpreters (#4930)
Problem

In some situations, the automatic python picker logic picks the wrong version of Python and we don't have much control over that beyond being able to specify version constraints (which sometimes aren't sufficient).

Solution

I added an option to explicitly specify a list of paths to search.

Result

Users can now pass in --python-setup-interpreter-search-paths=<search-paths> and have Pants search in the specified paths. You can even pass in the full path of the interpreter (e.g., which python) and it'll figure that out!