Reformat rust files (#4948)
Adds a pre-commit hook to ensure that they stay properly formatted.

I manually re-flowed some comments to make them <100 characters, because
we don't have wrap_comments on. I generally find auto-wrapping of
comments kind of sketchy, but we can turn it on if people needing to
re-flow comments manually becomes a big problem.