Use more generic portion of `requests` exception message in tests. (#4981)
### Problem

It appears that newer versions of `requests` (`2.18.4` vs `???`: argh, floating) result in slightly different error messages on linux. While we really shouldn't be floating that widely, it seems like it might be necessary to in order to allow pants to be built in contexts where other versions of requests are necessary.

An example failure: https://travis-ci.org/pantsbuild/pants/jobs/288657800

### Solution

Continue to float as begun with #4916, but use a more stable portion of the error message for comparison.

### Result

OSX and Linux both pass.