Reduce time/package requirements on binary-builder shards (#4992)
### Problem

The binary-builder shards currently run dry-run releases (using `-n`), which additionally runs the tests for contrib packages. But OSX time in travis continues to be a relatively scarce resource, and running the contrib tests can take up to 30 minutes in some cases. Additionally, running the contrib tests increases the number of tools required, all of which need to be installed in our (wheezy) docker image (see #4995).

### Solution

Add support to both test (`-t`) and dry-run (`-n`) releases for a "core only" `-c` flag, which skips testing contrib modules. Pass `-c` for both binary-builder shards.

Contrib continues to be tested on the dedicated contrib shards.

### Result

Both shards should run more quickly, and with fewer requirements.

Anecdotally, the difference with `-nc` on my laptop:
```
[=== 03:53 Dry run release succeeded ===]
```
vs a recent run of master in the OSX travis shard with `-n`:
```
[=== 32:48 Dry run release succeeded ===]
```