Add custom commands to the `buildozer` goal
This adds support for a `--command` option to the buildozer goal. This will be
particularly helpful for executing features beyond adding and removing
dependencies. For example, setting the name of an attribute, replacing an
attribute name, and printing attributes within a BUILD file are all custom
commands that can be executed now with --command.

Example to rename the name attribute to new_name in //tmp:tmp:

`./pants buildozer --command='set name new_name' //tmp:tmp`

This follows up on:
* https://github.com/pantsbuild/pants/pull/4921

This will also help out:
* https://github.com/pantsbuild/pants/pull/4882