Invalidate parent directories (#5000)
### Problem

Watchman `4.9.0` does not send events for a parent when a child is created or deleted, which causes issues like #4662.

### Solution

Revert to explicitly invalidating the parent directory when a child has changed. More precision is possible here, but it's possible that re-gaining the precision by using the `notify` crate would be preferable (see #4999).

### Result

Fixes #4662.