Working implementation of jacoco. (#4978)
### Problem

Jacoco is now available as an option, but the existing implementation is a no-op stub.

### Solution

Filled out the implementation of the jacoco code coverage engine.

### Result

Specifying jacoco as the code coverage processor now creates jacoco reports for junit tests. See the attachment for an example of the jacoco output against pants java code:

[coverage.zip](https://github.com/pantsbuild/pants/files/1385892/coverage.zip)

One place where I could use advice: what's the best way to handle working with the snapshot?

I'm using the snapshot because it was the only available version of jacoco that included the client tools. I could write a wrapper around the reporting classes, but it seemed best to avoid that if there was already a cli I could use, and there is...just not in any of the release versions. It's not clear when the next release will be (https://groups.google.com/forum/#!topic/jacoco/gd8xD30TDNo), so my feeling is that moving forward with the current snapshot build (as long as it seems to work) is the best bet. However, maybe pants has some space where I could push the current snapshot so we can guarantee a stable version until the next jacoco release hits maven central? Please let me know what seems best.