Move confluence related things to contrib (#4986)
### Problem
We want to move everything related to confluence to contrib instead of having to rely on the confluence module in `twitter.common` because we can't update it as freely. We also want to deprecate anything related to confluence in `pants.backend.docgen` because it doesn't make sense to have confluence targets and tasks live there. 

This change was initiated by a realization that the API the `twitter.common.confluence` module is using is [deprecated](https://developer.atlassian.com/confdev/deprecated-apis/confluence-xml-rpc-and-soap-apis) and we should switch over to the REST API. This change is required as a precursor to the API update.  

### Solution
Move anything confluence related to contrib, including the module from `twitter.common`.

Tested with:
`./pants test contrib/confluence/tests/python/pants_test/contrib/confluence/targets:wiki_page`
`./pants compile contrib/confluence/src/python/pants/contrib/confluence/tasks`