Fix memory leak in `./pants changed` (#5011)
### Problem

Before @kwlzn fixed #4508 by introducing `compute_dependency_specs`, there was no way to get the "implicit" dependencies of a Target without constructing it. So #4424 switched to using a full `BuildGraph` to compute the dependents in changed. While that fixed the bug, it also introduced a memory leak, because instantiating a Target causes a cycle with the BuildGraph that owns it. 

### Solution

Now that #4508 is fixed, restore the codepaths that were replaced in #4424, and use `compute_dependency_specs` to compute the "implicit" dependencies of Targets in the graph. Additionally, cleans up the `tool.jar`/`javac` split in the Java subsystem to remove the requirement that `def injectables` has been called before requesting injectable specs.

### Result

Fixes #4494.