Use the service deps if the target declares an exception. (#5017)
### Problem

Thrift exceptions require the same dependencies declared by services (namely they require com.twitter.finagle.SourcedException, which comes in from finagle), but are frequently used by callers similarly to structs.

scrooge and pants provide different sets of dependencies to structs-only targets than they do to targets which contain services. To resolve this, we would either need to split the classification again, into: "declares structs", "declares exceptions", and "declares services", or do a union based on which symbols are found in the file.

For now, the workaround is to add an empty service to a target that needs to use exceptions, or to move the exceptions to a target that defines a service instead.

### Solution

For sources that declare services OR exceptions, we now use the service deps.

### Result

Users will no longer need to use empty services or move the exceptions to targets that declare services.