Fix JarCreate invalidation in the presence of changing resources. (#5030)
### Problem

As described in #4595, `JarCreate` caches jar contents too aggressively. The root cause is that the `JarTask` superclass looks at "some of the dependencies of a target" (in the case of that particular ticket: resources), but the target invalidation in `JarCreate` is configured to only invalidate for changes directly in the target.

This is another case of the "whitelist based cache key" issue, and is one of the primary motivators for moving to isolated-process-executions as cache entries.

### Solution

As described in the change: rather than attempting to whitelist exactly what `JarTask` touches, turning on `invalidate_dependents` is more durable in the face of further code changes. We do that here.

### Result

Fixes #4595.