Add configurable message when missing-deps-suggest doesn't have suggestions (#5036)
Problem

It would be good to have a configurable message that will be printed when missing-deps-suggest cannot find any targets. The goal of this message is to point people to some documentation that might help solve their dependency problems.

Solution

Add an option that takes a string and logs it as an error when missing-deps-suggest matches a class not found error but doesn't have any suggestions. Please not that I upgraded the existing logs from debug to error because I thought it was useful to see the exact classes that couldn't be matched.