Fixup `./build-support/bin/release.sh -t`.
Previously the install and test function for `pantsbuild.pants` and
`pantsbuild.pants.testinfra` pip-installed a local wheel instead of
letting the pip enviornment determine whether to find a local wheel or
try to fetch a distribution. Fix this so that these two packages are
asked for by version requirement only just like all the other (contrib)
packages.