Add a VFS impl for PosixFS. (#5079)
### Problem

`PosixFS` cannot be used directly to expand `PathGlobs` into `PathStats`, but that would be very useful for synchronous/non-incremental snapshot capture.

### Solution

Add a `VFS` impl for `PosixFS`. Given further tweaks to align the parameters types and the self type (not worthwhile now?) the implementations of the methods could likely move into the VFS impl, rather than staying in `impl PosixFS` and needing to be proxied to.