Fix `PythonInterpreterCache`.
Previously, `PythonInterpreterCache` was broken in several ways:

1. Use was delicate: `setup` had to be called before
`select_interpreter_for_targets` and an appropriate set of filters
derived from the targets passed to `select_interpreter_for_targets`
needed to be passed to `setup` to ensure a compatible
`PythonInterpreter` was selected when one was available.
2. Locking was left to the caller.

In practice, Pants code did not pass filters to `setup` and so was
broken by 1 (see: #5081) and the locking done was via a Pants workspace
file lock, which was broken for `PythonSetup`s that used a global
`--resolver-cache-dir` (which was the case with the Pants repo itself).

Fix `PythonInterpreterCache` by removing the requirement to call `setup`
first: `select_interpreter_for_targets` now derives the correct filter
set from the targets it's given and calls setup with these. Also move the
locking responsibility into `PythonInterpreterCache`, ensuring safe
concurrent usage regardless of context.

Fixes #5081