Inline mis-documented `select_interpreter` method. (#5085)
The `select_interpreter` method claimed to be picking a _compatible_
interpreter but it relied on context to do so; as such, a compatible
interpreter was never being selected for calling code in the `Export`
task. Simplify by inlining the logic and adding a hefty note about the
suspect choice Export has been making.