Target Rename with Buildozer  (#5023)
### Problem
What if an owner of a popular library wants to change the Pants target name for that library? Then they would have to update multiple BUILD files to ensure that all of the library's dependees use the new name.

### Solution
The solution adds a pants goal called 'meta-rename' to the `buildrefactor` module. This allows us to update the name of a target with --from and --to options. The solution itself builds a dependency graph of the repository. Then by using buildozer, we can rename the original build file, and rename wherever it's referenced as a dependency.